Isolate type evolution code into a new module
